Spelling Update DataContractAttribute.cs and DataMemberMode.cs
docs(api): Correct minor spelling and grammar.

There were instances of the words 'property / field' 'property or field' and 'field/property' so I chose 'or' and standardized all occurrences.
Some double spaces, or extra words.

The DataMemberMode property summary uses more precise language to include structs. Unified the details to the DataContractAttribute.

Summary 'Gets or sets' is unnecessary in summary. More to the point and maintainable to remove.

REFACTOR: 'serialized' language is used. However, everywhere else 'store' and 'restore' are used. Can we unify this?